fix: Moved the cyclehiredockingdata object
