--- app/graphql/types/cycle_hire_docking_station_data_type.rb
@@ -1,10 +1,0 @@
-module Types
-  class CycleHireDockingStationDataType < Types::BaseObject
-    field :destinationId, String, null: true
-    field :destinationNumberOfBikes, Integer, null: true
-    field :destinationNumberOfEmptySlots, Integer, null: true
-    field :originId, String, null: true
-    field :originNumberOfBikes, Integer, null: true
-    field :originNumberOfEmptySlots, Integer, null: true
-  end
-end
