--- app/graphql/types/journey_search_result_type.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  class JourneyPlannerResultType < Types::BaseObject
-    field :recommendedMaxAgeMinutes, Integer, null: true
-  end
-end
