+++ spec/graphql/types/tfl/entities/journey_planner/cycle_hire_docking_station_data_type_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Types::CycleHireDockingStationDataType do
+describe Types::Tfl::Entities::JourneyPlanner::CycleHireDockingStationDataType do
   subject { described_class }
 
   it { is_expected.to have_field(:destinationId).of_type('String') }
